chore(virtctl): Move user list command to own file
