+++ pkg/virtctl/vm/vm.go
@@ -20,8 +20,6 @@
 package vm
 
 import (
-	"context"
-	"encoding/json"
 	"fmt"
 	"strings"
 
@@ -31,13 +29,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"kubevirt.io/client-go/kubecli"
-
-	"kubevirt.io/kubevirt/pkg/virtctl/templates"
 )
 
 const (
-	COMMAND_USERLIST = "userlist"
-
 	volumeNameArg         = "volume-name"
 	notDefinedGracePeriod = -1
 	dryRunCommandUsage    = "--dry-run=false: Flag used to set whether to perform a dry run or not. If true the command will be executed without performing any changes."
@@ -59,21 +53,6 @@ var (
 	persist      bool
 	dryRun       bool
 )
-
-func NewUserListCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:     "userlist (VMI)",
-		Short:   "Return full list of logged in users on the guest machine.",
-		Example: usage(COMMAND_USERLIST),
-		Args:    templates.ExactArgs("userlist", 1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			c := Command{command: COMMAND_USERLIST, clientConfig: clientConfig}
-			return c.Run(args)
-		},
-	}
-	cmd.SetUsageTemplate(templates.UsageTemplate())
-	return cmd
-}
 
 type Command struct {
 	clientConfig clientcmd.ClientConfig
@@ -122,21 +101,6 @@ func (o *Command) Run(args []string) error {
 	dryRunOption := getDryRunOption(dryRun)
 	if len(dryRunOption) > 0 && dryRunOption[0] == metav1.DryRunAll {
 		fmt.Printf("Dry Run execution\n")
-	}
-	switch o.command {
-	case COMMAND_USERLIST:
-		userlist, err := virtClient.VirtualMachineInstance(namespace).UserList(context.Background(), vmiName)
-		if err != nil {
-			return fmt.Errorf("Error listing users of VirtualMachineInstance %s, %v", vmiName, err)
-		}
-
-		data, err := json.MarshalIndent(userlist, "", "  ")
-		if err != nil {
-			return fmt.Errorf("Cannot marshal userlist %v", err)
-		}
-
-		fmt.Printf("%s\n", string(data))
-		return nil
 	}
 
 	fmt.Printf("VM %s was scheduled to %s\n", vmiName, o.command)
